+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Added: support for social sync bundle
 - Added: information if no items found to social stats command
-- 
 
 ## [1.5.6] - 2018-12-06